FEAT #11269 TIME 1:15 add control + add history (indexing Model and indexing field)

+++ src/app/customField/controllers/CustomFieldController.php
@@ -19,6 +19,7 @@ namespace CustomField\controllers;
 
 use CustomField\models\CustomFieldModel;
 use Group\models\ServiceModel;
+use History\controllers\HistoryController;
 use IndexingModel\models\IndexingModelFieldModel;
 use Respect\Validation\Validator;
 use Slim\Http\Request;
@@ -64,6 +65,15 @@ class CustomFieldController
             'values'        => empty($body['values']) ? '[]' : json_encode($body['values'])
         ]);
 
+        HistoryController::add([
+            'tableName' => 'custom_fields',
+            'recordId'  => $id,
+            'eventType' => 'ADD',
+            'info'      => _CUSTOMFIELDS_CREATION . " : {$body['label']}",
+            'moduleId'  => 'customField',
+            'eventId'   => 'customFieldCreation',
+        ]);
+
         return $response->withStatus(201)->withJson(['customFieldId' => $id]);
     }
 
@@ -73,9 +83,8 @@ class CustomFieldController
             return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
         }
 
-        $field = CustomFieldModel::getById(['select' => [1], 'id' => $args['id']]);
-        if (empty($field)) {
-            return $response->withStatus(400)->withJson(['errors' => 'Custom field not found']);
+        if (!Validator::intType()->notEmpty()->validate($args['id'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Param id is empty or not an integer']);
         }
 
         $body = $request->getParsedBody();
@@ -90,6 +99,11 @@ class CustomFieldController
             return $response->withStatus(400)->withJson(['errors' => 'Some values have the same name']);
         }
 
+        $field = CustomFieldModel::getById(['select' => [1], 'id' => $args['id']]);
+        if (empty($field)) {
+            return $response->withStatus(400)->withJson(['errors' => 'Custom field not found']);
+        }
+
         $fields = CustomFieldModel::get(['select' => [1], 'where' => ['label = ?', 'id != ?'], 'data' => [$body['label'], $args['id']]]);
         if (!empty($fields)) {
             return $response->withStatus(400)->withJson(['errors' => 'Custom field with this label already exists']);
@@ -97,11 +111,20 @@ class CustomFieldController
 
         CustomFieldModel::update([
             'set'   => [
-                'label'         => $body['label'],
-                'values'        => empty($body['values']) ? '[]' : json_encode($body['values'])
+                'label'  => $body['label'],
+                'values' => empty($body['values']) ? '[]' : json_encode($body['values'])
             ],
             'where' => ['id = ?'],
             'data'  => [$args['id']]
+        ]);
+
+        HistoryController::add([
+            'tableName' => 'custom_fields',
+            'recordId'  => $args['id'],
+            'eventType' => 'UP',
+            'info'      => _CUSTOMFIELDS_MODIFICATION . " : {$body['label']}",
+            'moduleId'  => 'customField',
+            'eventId'   => 'customFieldModification',
         ]);
 
         return $response->withStatus(204);
@@ -113,6 +136,12 @@ class CustomFieldController
             return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
         }
 
+        if (!Validator::intType()->notEmpty()->validate($args['id'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Param id is empty or not an integer']);
+        }
+
+        $field = CustomFieldModel::getById(['select' => ['label'], 'id' => $args['id']]);
+
         IndexingModelFieldModel::delete(['where' => ['identifier = ?'], 'data' => [$args['id']]]);
 
         //TODO Suppression des valeurs liés aux courriers ?
@@ -120,6 +149,15 @@ class CustomFieldController
         CustomFieldModel::delete([
             'where' => ['id = ?'],
             'data'  => [$args['id']]
+        ]);
+
+        HistoryController::add([
+            'tableName' => 'custom_fields',
+            'recordId'  => $args['id'],
+            'eventType' => 'DEL',
+            'info'      => _CUSTOMFIELDS_SUPPRESSION . " : {$field['label']}",
+            'moduleId'  => 'customField',
+            'eventId'   => 'customFieldSuppression',
         ]);
 
         return $response->withStatus(204);
